requirements.txt
Flask~=2.0.2
flask-smorest~=0.35.0
pyctuator~=0.17.0
pycryptodome~=3.11.0